requirements.txt
asgiref==3.8.1
attrs==25.1.0
autopep8==2.3.2
certifi==2025.1.31
distlib==0.3.9
Django==5.1.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
filelock==3.17.0
gprof2dot==2024.6.6
gunicorn==23.0.0
inflection==0.5.1
isort==6.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pillow==10.4.0
pipenv==2024.4.1
platformdirs==4.3.6
pycodestyle==2.12.1
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
setuptools==75.8.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.29.1
waitress==3.0.2
